requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
facenet-pytorch>=2.5.3
albumentations>=1.3.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0
tensorboard>=2.15.0
streamlit>=1.28.0
